% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_egflow.R
\name{use_gitignore_egflow}
\alias{use_gitignore_egflow}
\title{use_gitignore}
\usage{
use_gitignore_egflow()
}
\value{
nothing, creates a file.
}
\description{
Use a starter .gitignore
}
\details{
Drop a starter .gitignore in the current working directory, including
ignores for targets and capsule (renv).
}
\author{
Eric Gaudiello
}
