% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_egflow.R
\name{use_project_egflow}
\alias{use_project_egflow}
\title{use_project_egflow}
\usage{
use_project_egflow()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup an egflow default project
}
\details{
Creates files and directories according to the egflow project template.
}
\author{
Eric Gaudiello
}
